added Qing Vector Map
